// src/utils/enums.ts
export const enum Label {
  Heading = 'Heading',
  Subheading = 'Subheading',
  Thumbnail = 'Thumbnail',
  Standfirst = 'Standfirst',
  Author = 'Author',
  PubDate = 'Publication date',
  Body = 'Body',
  Categories = 'Category',
  Tags = 'Tag',
  Links = 'Link',
  Images = 'Image',
  Summary = 'Summary',
  BodyExclusion = 'Body exclusion',
}

export const enum ErrorName {
  QuerySelector = 'QuerySelectorError',
  UnsupportedError = 'UnsupportedError',
}